Add constant for sign-up agreement message
This commit adds a new constant class AgreementMessage to the project. This class defines a constant value signUpAgreementMessage that represents the message displayed in the sign-up agreement.

The sign-up agreement message emphasizes the importance of privacy and assures users that any personal information they provide, such as their email address and contact number, will be protected in accordance with the Data Privacy Act of 2012. By submitting their information, users give their full and voluntary consent to the collection of this data.

The signUpAgreementMessage constant is defined as a String and contains the specific text of the message. It is designed to be easily accessible and reusable throughout the application.

By adding this constant, the code enhances the clarity and maintainability of the sign-up feature by centralizing the message text in one place. It ensures consistency in the wording of the sign-up agreement message and simplifies any future updates or modifications to the message.