chore: Update main function to remove URL number sign and add comments
This commit updates the main function in the project. The changes include:

Importing the url_strategy package to use the setPathUrlStrategy function.
Calling setPathUrlStrategy() to remove the number sign (#) from the URL.
Adding comments to explain the purpose of the code.
By removing the number sign from the URL, the application's URLs will have a cleaner appearance and a more user-friendly structure.