feat: Add cloud_firestore and animated_text_kit dependencies to pubspec.yaml
This commit adds the cloud_firestore and animated_text_kit dependencies to the pubspec.yaml file.

The cloud_firestore package is added to enable integration with Firestore, a flexible and scalable NoSQL cloud database provided by Firebase. This dependency allows for efficient storage and retrieval of data in the application.

The animated_text_kit package is included to provide animated text effects and transitions in the user interface. This package offers various animation styles and effects that can be applied to text elements, enhancing the visual appeal and user experience of the application.

By adding these dependencies to the pubspec.yaml file, the project gains access to the functionality provided by the cloud_firestore and animated_text_kit packages, enabling seamless integration with Firestore and the implementation of captivating text animations within the application.