feat: add text animation while waiting on create account
The code defines the creatingAccountTextAnimation function, which displays a dialog with animated text to provide visual feedback during the account creation process.

Here's how the function works:

The creatingAccountTextAnimation function takes two parameters:

context: The build context of the widget.
themeState: The current theme state, which is used to determine the text color based on different theme options (dark theme, mirage theme, or other light themes).
The function creates a dialog using the showDialog method. The barrierDismissible property is set to false to prevent dismissing the dialog by tapping outside.

Inside the dialog, a Dialog widget is used with a transparent background color (Colors.transparent) and zero elevation (elevation: 0) to achieve a custom dialog appearance.

The animated text effect is achieved using the AnimatedTextKit widget from the animated_text_kit package. It allows displaying a series of animated texts with various effects.

The DefaultTextStyle widget sets the default text style for the animated text. The text size is set to 30, and the color is determined based on the themeState parameter. For dark theme, the text color is white; for mirage theme, it's white70; and for other light themes, it's based on the scaffold background color from the LightThemeColors class.

The animated texts are defined as a list of RotateAnimatedText widgets, each representing a text phrase. In this case, the phrases are "CREATING YOUR ACCOUNT," "WE APPRICIATE YOUR PATIENCE," and "YOU ARE AMAZING." These texts will be displayed sequentially in the dialog with a rotating animation and fade-in effect.

The function wraps the entire dialog creation code and returns a Future<dynamic>, which allows it to be awaited for further control flow if needed.