feat: add dialog box button
The code defines the DialogBoxButton widget, which represents a button commonly used within dialog boxes. Here's how the widget works:

The DialogBoxButton widget is a stateless widget that takes three parameters:

themeState: The current theme state, which is used to determine the button style based on the elevatedButtonStyle function from the button_style.dart file.
onPressed: An optional callback function that is executed when the button is pressed. If not provided, the default action is to dismiss the dialog by calling context.pop().
label: An optional string that represents the text displayed on the button. If not provided, the default label is "Okay".
The widget returns a SizedBox containing an ElevatedButton widget with the specified properties.

The SizedBox is used to set the width and height of the button to 100 and 30, respectively. You can modify these values to adjust the button's size according to your requirements.

The ElevatedButton widget is used to create the button. It uses the elevatedButtonStyle function from the button_style.dart file to determine the button's visual style based on the themeState parameter.

The onPressed property of the ElevatedButton widget is set to the provided onPressed callback function. If no onPressed function is provided, the default action is to dismiss the dialog by calling context.pop().

The child property of the ElevatedButton widget is set to a Text widget that displays the button label. The label is either the provided label string or the default label "Okay".

The DialogBoxButton widget simplifies the creation of action buttons within dialog boxes by encapsulating the button's appearance and behavior. It allows for easy customization and reuse in different dialog scenarios.